+++ 2025/day2/Range.h
@@ -1,11 +1,12 @@
 #pragma once
 
+#include <cstdint>
 #include <string>
 
 struct Range
 {
-    size_t begin;
-    size_t end;
+    uint32_t begin;
+    uint32_t end;
 };
 
 Range ParseRange(const std::string& line);
